Created the attack game screen. This will be where the actual gameplay occurs.
Currently, the attack screen renders the player
and the computers grid and displays them.

Also made it so the player is able to
click on a square to see if they hit a ship.
A white circle shows if a miss occurs and a red
circle shows if a hit occured.